Change student_course grade column from integer to float

+++ db/migrate/20191103021046_add_grade_to_student_courses.rb
@@ -1,5 +1,5 @@
 class AddGradeToStudentCourses < ActiveRecord::Migration[5.1]
   def change
-    add_column :student_courses, :grade, :integer
+    add_column :student_courses, :grade, :float
   end
 end
